Transpose commit for global_histogram
